+++ ckb-indexer/src/main/java/org/nervos/indexer/model/Filter.java
@@ -1,7 +1,6 @@
 package org.nervos.indexer.model;
 
 import com.google.gson.annotations.SerializedName;
-import java.math.BigInteger;
 import java.util.List;
 import org.nervos.ckb.type.Script;
 
@@ -12,7 +11,7 @@ public class Filter {
   public List<Integer> outputDataLenRange;
 
   @SerializedName("output_capacity_range")
-  public List<BigInteger> outputCapacityRange;
+  public List<Long> outputCapacityRange;
 
   @SerializedName("block_range")
   public List<Integer> blockRange;
