+++ ckb-indexer/src/main/java/org/nervos/indexer/model/resp/CellCapacityResponse.java
@@ -10,5 +10,5 @@ public class CellCapacityResponse {
   @SerializedName("block_number")
   public int blockNumber;
 
-  public BigInteger capacity;
+  public long capacity;
 }
