+++ ckb-indexer/src/main/java/org/nervos/indexer/model/resp/OutputResponse.java
@@ -1,10 +1,9 @@
 package org.nervos.indexer.model.resp;
 
-import java.math.BigInteger;
 import org.nervos.ckb.type.Script;
 
 public class OutputResponse {
-  public BigInteger capacity;
+  public long capacity;
   public Script lock;
   public Script type;
 }
