+++ example/src/main/java/org/nervos/ckb/indexer/Receiver.java
@@ -5,9 +5,9 @@ import java.math.BigInteger;
 /** Copyright © 2019 Nervos Foundation. All rights reserved. */
 public class Receiver {
   public String address;
-  public BigInteger capacity;
+  public long capacity;
 
-  public Receiver(String address, BigInteger capacity) {
+  public Receiver(String address, long capacity) {
     this.address = address;
     this.capacity = capacity;
   }
